requirements.txt
streamlit==1.32.0
python-dotenv==1.0.1
requests>=2.32.2
pandas==2.2.1
altair==5.2.0 